Create loops1
In this challenge, you will use a for loop to increment a variable through a range.

Input Format

You will be given two positive integers, a and b  separated by a newline.

Output Format

For each integer n in the inclusive interval : [a,b]

If 1<=n<=9then print the English representation of it in lowercase. That is "one" for , "two" for , and so on.
Else if n>9 and it is an even number, then print "even".
Else if n>9 and it is an odd number, then print "odd".